+++ NEWS.md
@@ -14,6 +14,8 @@
 
 * `cb_summarize_categorical()` no longer throws an error when a variable contains no value labels.
 
+* `cb_create()` no longer throws an error when metadata when `.val_labels = NULL` (fixes #34).
+
 ## Internal
 
 * Added a test suite.